# src/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8082",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/batches/{bathId}/temperature": {
            "post": {
                "tags": [
                    "Yogurt Batch"
                ],
                "summary": "Registrar temperatura",
                "description": "Registra una medicion de temperatura en el lote",
                "operationId": "recordTemEntity",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TemperatureRecord"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Lote no encontrado"
                    },
                    "200": {
                        "description": "Temperatura registrada"
                    }
                }
            }
        },
...
        "/api/monitoring/batches/{batchId}/temperature": {
            "get": {
                "tags": [
                    "Monitoring"
                ],
                "summary": "Resumen de temperatura",
                "description": "Obtiene estadísticas de temperatura de un lote",
                "operationId": "getBatchTemperatureSummary",
                "parameters": [
                    {
                        "name": "batchId",
                        "in": "path",
                        "description": "ID del lote",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resumen obtenido",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemperatureSummary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lote no encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemperatureSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/monitoring/batches/{batchId}/temperature-logs": {
            "get": {
                "tags": [
                    "Monitoring"
                ],
                "summary": "Historial de temperaturas",
                "description": "Obtiene registros de temperatura de un lote, opcionalmente filtrados por rango de fechas",
                "operationId": "getTemperatureLogs",
                "parameters": [
                    {
                        "name": "batchId",
                        "in": "path",
                        "description": "ID del lote",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Fecha inicio (ISO 8601)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "example": "2026-04-27T10:00:00"
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Fecha fin (ISO 8601)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "example": "2026-04-27T12:00:00"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de registros obtenida",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TemperatureLog"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Ingredient": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Nombre del ingrediente",
                        "example": "Leche entera"
                    },
                    "quantity": {
                        "type": "number",
                        "description": "Cantidad del ingrediente",
                        "format": "double",
                        "example": 1.0
                    },
                    "unit": {
                        "type": "string",
                        "description": "Unidad de medida",
                        "example": "litros"
                    },
                    "notes": {
                        "type": "string",
                        "description": "Notas adicionales sobre el ingrediente",
                        "example": "Preferiblemente fresca"
                    },
                    "options": {
                        "type": "boolean",
                        "description": "Indica si el ingrediente es opcional",
                        "example": false
                    }
                },
                "description": "Ingrediente utilizado en una receta de yogurt"
            },
            "Recipe": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Nombre de la receta",
                        "example": "Yogurt natural clásico"
                    },
                    "description": {
                        "type": "string",
                        "description": "Descripción de la receta",
                        "example": "Receta tradicional de yogurt casero"
                    },
                    "defaulMilkVolume": {
                        "type": "number",
                        "description": "Volumen de leche en litros",
                        "format": "double",
                        "example": 1.0
                    },
                    "defaulStarterAmount": {
                        "type": "number",
                        "description": "Cantidad de cultivo iniciador (starter) en gramos",
                        "format": "double",
                        "example": 50.0
                    },
                    "heatingTemperature": {
                        "type": "number",
                        "description": "Temperatura de calentamiento en °C",
                        "format": "double",
                        "example": 85.0
                    },
                    "heatingDuration": {
                        "type": "integer",
                        "description": "Duración del calentamiento en minutos",
                        "format": "int32",
                        "example": 30
                    },
                    "innoculationTemperature": {
                        "type": "number",
                        "description": "Temperatura de inoculación en °C",
                        "format": "double",
                        "example": 43.0
                    },
                    "incubationTemperature": {
                        "type": "number",
                        "description": "Temperatura de incubación en °C",
                        "format": "double",
                        "example": 42.0
                    },
                    "minIncubationTime": {
                        "type": "integer",
                        "description": "Tiempo mínimo de incubación en minutos",
                        "format": "int32",
                        "example": 240
                    },
                    "maxIncubationTime": {
                        "type": "integer",
                        "description": "Tiempo máximo de incubación en minutos",
                        "format": "int32",
                        "example": 480
                    },
                    "refrigerationTime": {
                        "type": "integer",
                        "description": "Tiempo de refrigeración en minutos",
                        "format": "int32",
                        "example": 180
                    },
                    "difficulty": {
                        "type": "string",
                        "description": "Nivel de dificultad de la receta",
                        "example": "BEGINNER",
                        "enum": [
                            "BEGINNER",
                            "INTERMEDIATE",
                            "ADVANCED"
                        ]
                    },
                    "tips": {
                        "type": "string",
                        "description": "Consejos adicionales",
                        "example": "Mantener temperatura constante durante la incubación"
                    },
                    "ingredients": {
                        "type": "array",
                        "description": "Lista de ingredientes de la receta",
                        "items": {
                            "$ref": "#/components/schemas/Ingredient"
                        }
                    }
                },
                "description": "DTO para la creación y actualización de recetas de yogurt"
            },
...
            "TemperatureLog": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "batch": {
                        "$ref": "#/components/schemas/YogurtBatch"
                    },
                    "temperature": {
                        "type": "number",
                        "format": "double"
                    },
                    "recordedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "HEATING",
                            "COOLING",
                            "INCUBATION",
                            "REFRIGERATION",
                            "MANUAL"
                        ]
                    },
                    "notes": {
                        "type": "string"
                    }
                }
            },
            "YogurtBatch": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "batchCode": {
                        "type": "string"
                    },
                    "recipe": {
                        "$ref": "#/components/schemas/Recipe"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PREPARING",
                            "HEATING",
                            "COOLING",
                            "INOCULATING",
                            "INCUBATING",
                            "REFRIGERATING",
                            "COMPLETED",
                            "FAILED"
                        ]
                    },
                    "milkVolume": {
                        "type": "number",
                        "format": "double"
                    },
                    "starterAmount": {
                        "type": "number",
                        "format": "double"
                    },
                    "targetTemperature": {
                        "type": "number",
                        "format": "double"
                    },
                    "incubationTime": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "startTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "incubationStartTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "incubationEndTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "refrigerationStartTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "temperatureLogs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TemperatureLog"
                        }
                    },
                    "notes": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "TemperatureRecord": {
                "required": [
                    "temperature"
                ],
                "type": "object",
                "properties": {
                    "temperature": {
                        "type": "number",
                        "description": "Temperatura registrada en °C",
                        "format": "double",
                        "example": 42.5
                    },
                    "type": {
                        "type": "string",
                        "description": "Tipo de registro de temperatura",
                        "example": "INCUBATION",
                        "enum": [
                            "HEATING",
                            "COOLING",
                            "INCUBATION",
                            "REFRIGERATION",
                            "MANUAL",
                            "HEATING",
                            "COOLING",
                            "INCUBATION",
                            "REFRIGERATION",
                            "MANUAL"
                        ]
                    }
                },
                "description": "Registro de temperatura para un lote de yogurt"
            },
...
            "TemperatureSummary": {
                "type": "object",
                "properties": {
                    "currentTemperature": {
                        "type": "number",
                        "description": "Temperatura actual del lote en °C",
                        "format": "double",
                        "example": 42.5
                    },
                    "maximunTemperature": {
                        "type": "number",
                        "description": "Temperatura máxima registrada en °C",
                        "format": "double",
                        "example": 45.0
                    },
                    "minmumTemperature": {
                        "type": "number",
                        "description": "Temperatura mínima registrada en °C",
                        "format": "double",
                        "example": 38.5
                    },
                    "averageTemperature": {
                        "type": "number",
                        "description": "Temperatura promedio durante la incubación en °C",
                        "format": "double",
                        "example": 42.0
                    }
                },
                "description": "Resumen de estadísticas de temperatura de un lote"
            }
        }
    }
}